backend/requirements.txt
fastapi
uvicorn
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
sqlalchemy
python-multipart
pillow
python-dotenv